requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Limiter==3.5.0
WTForms==3.1.1
Werkzeug==3.0.1
python-dateutil==2.8.2
firebase-admin==6.4.0
gunicorn==21.2.0
